restructure fitting code
