+++ src/tests/unittests/algorithms/test_fitting.py
@@ -4,8 +4,8 @@ import unittest
 import numpy as np
 import matplotlib.pyplot as plt
 import qtt
-from qtt.algorithms.functions import FermiLinear, linear_function, double_gaussian
-from qtt.algorithms.fitting import initFermiLinear, _estimate_fermi_model_center_amplitude, fitFermiLinear, fit_double_gaussian, refit_double_gaussian
+from qtt.algorithms.functions import FermiLinear, linear_function, double_gaussian, gaussian
+from qtt.algorithms.fitting import initFermiLinear, _estimate_fermi_model_center_amplitude, fitFermiLinear, fit_double_gaussian, refit_double_gaussian, fit_gaussian
 
 
 class TestDoubleGaussianFitting(unittest.TestCase):
@@ -31,6 +31,27 @@ class TestDoubleGaussianFitting(unittest.TestCase):
         fitted_parameters, _ = fit_double_gaussian(x_data, y_data)
         parameter_diff = np.abs(fitted_parameters - initial_parameters)
         self.assertTrue(np.all(parameter_diff < 1e-3))
+
+class TestGaussianFitting(unittest.TestCase):
+
+    def test_fit_gaussian_no_offset(self):
+        x_data = np.linspace(0, 10, 100)
+        gauss_data = gaussian(x_data, mean=4, std=1, amplitude=5)
+        noise = np.random.rand(100) - .5
+        result_dict = qtt.algorithms.functions.fit_gaussian(x_data=x_data, y_data=(gauss_data + noise), estimate_offset=0)
+        np.testing.assert_array_almost_equal(result_dict['fitted_parameters'], np.array([4, 1, 5.]), decimal=1)
+        self.assertTrue(result_dict['reduced_chi_squared']< .2)
+
+    def test_fit_gaussian(self):
+        x_data = np.linspace(0, 10, 100)
+        gauss_data = 0.1 + gaussian(x_data, mean=4, std=1, amplitude=5)
+        noise = np.random.rand(100)-.5
+        [mean, s, amplitude, offset], _ = fit_gaussian(x_data=x_data, y_data=(gauss_data + noise))
+        self.assertTrue(3.5 < mean < 4.5)
+        self.assertTrue(0.5 < s < 1.5)
+        self.assertTrue(4.5 < amplitude < 5.5)
+        self.assertAlmostEqual(offset, 0.1, places = 0)
+
 
 class TestFermiFitting(unittest.TestCase):
